feat: 스크롤 이벤트 커스텀 훅으로 변경
